fix shit

+++ src/engine/instrument.cpp
@@ -757,7 +757,7 @@ void DivInstrument::writeFeatureNE(SafeWriter* w) {
 void DivInstrument::writeFeatureE3(SafeWriter* w) {
   FEATURE_BEGIN("E3");
   
-  w->writeC(es5503.wavePos);
+  w->writeC((uint8_t)es5503.wavePos);
   w->writeC((es5503.initial_osc_mode << 6)|(es5503.softpan_virtual_channel << 5)|(es5503.phase_reset_on_start << 4)|(es5503.waveLen << 1)|(es5503.auto_place_wavetables));
 
   FEATURE_END;
@@ -2631,7 +2631,7 @@ void DivInstrument::readFeatureNE(SafeReader& reader, short version) {
 void DivInstrument::readFeatureE3(SafeReader& reader, short version) {
   READ_FEAT_BEGIN;
 
-  es5503.wavePos=reader.readC();
+  es5503.wavePos=(uint8_t)reader.readC();
   uint8_t temp = reader.readC();
 
   es5503.initial_osc_mode = temp >> 6;
